Move common optim_fun to another file
